+++ src/compilation/ir/instructions/PutInstruction.java
@@ -2,7 +2,7 @@ package compilation.ir.instructions;
 
 import compilation.ir.UniqueVariable;
 
-public record PutInstruction(UniqueVariable variable, byte value) implements Instruction {
+public record PutInstruction(UniqueVariable variable, int value) implements Instruction {
     @Override
     public <R> R accept(InstructionVisitor<R> visitor) {
         return visitor.visitPutInstruction(this);
